requirements.txt
absl-py==2.0.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
cloudpickle==3.0.0
cmake==3.27.7
contextlib2==21.6.0
contourpy==1.2.0
cubinlinker-cu11==0.3.0.post1
cuda-python==11.8.2
cudf-cu11==23.10.2
cuml-cu11==23.10.0
cupy-cuda11x==12.2.0
cycler==0.12.1
Cython==3.0.5
dask==2023.9.2
dask-cuda==23.10.0
dask-cudf-cu11==23.10.2
dinov2 @ git+https://github.com/facebookresearch/dinov2@da4b3825f0ed64b7398ace00c5062503811d0cff
distributed==2023.9.2
docker-pycreds==0.4.0
easydict==1.11
einops==0.7.0
fastrlock==0.8.2
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
ftfy==6.1.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.3
huggingface-hub==0.19.4
idna==3.4
imageio==2.33.0
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lanms-neo==1.0.2
lazy_loader==0.3
lit==17.0.5
llvmlite==0.40.1
lmdb==1.4.1
locket==1.0.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-collections==0.1.1
mmcv==1.7.1
mmdet==2.28.2
mmengine==0.10.1
mmocr==0.6.3
mpmath==1.3.0
msgpack==1.0.7
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.57.1
numpy==1.24.4
nvtx==0.2.8
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==23.2
pandas==1.5.3
partd==1.4.1
Pillow==10.1.0
platformdirs==4.0.0
portalocker==2.8.2
protobuf==4.23.4
psutil==5.9.6
ptflops==0.7.1.2
ptxcompiler-cu11==0.7.0.post1
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
Pygments==2.17.2
pylibraft-cu11==23.10.0
pynvml==11.4.1
pyparsing==3.1.1
pyre-extensions==0.0.23
python-dateutil==2.8.2
pytorch-toolbelt==0.6.3
pytz==2023.3.post1
PyYAML==6.0.1
raft-dask-cu11==23.10.0
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rmm-cu11==23.10.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
sentry-sdk==1.36.0
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
submitit==1.5.1
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.3.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0+cu117
torchmetrics==0.10.3
torchvision==0.15.0+cu117
tornado==6.3.3
tqdm==4.66.1
treelite==3.9.1
treelite-runtime==3.9.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.8.0
ucx-py-cu11==0.34.0
urllib3==2.1.0
wandb==0.16.0
wcwidth==0.2.12
Werkzeug==3.0.1
xformers==0.0.18
yacs==0.1.8
yapf==0.40.2
zict==3.0.0
zipp==3.17.0
